# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exe Joiner API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  discordId: string;
  username: string;
  avatar?: string | null;
  email?: string | null;
  isAdmin: boolean;
}

export interface Slot {
  id: number;
  slotNumber: number;
  isActive: boolean;
  purchasedAt?: string | null;
  expiresAt?: string | null;
  label?: string | null;
}

export interface SlotsResponse {
  slots: Slot[];
  totalSlots: number;
  pricePerDay: number;
}

export interface AdminSettings {
  slotCount: number;
  pricePerDay: number;
}

export interface AdminSettingsUpdate {
  /**
   * @minimum 1
   * @maximum 100
   */
  slotCount?: number;
  /** @minimum 0 */
  pricePerDay?: number;
}

export interface AdminUserEntry {
  discordId: string;
  username: string;
  avatar?: string | null;
  activeSlots: number;
  totalSlots: number;
}

export interface AdminUsersResponse {
  users: AdminUserEntry[];
}

export interface AdminUpdateUserSlotsRequest {
  /**
   * @minimum 0
   * @maximum 100
   */
  activeSlotCount: number;
}

export interface CreateStripeSessionRequest {
  /**
   * @minimum 1
   * @maximum 100
   */
  slotNumber: number;
}

export interface StripeSessionResponse {
  sessionId: string;
  url: string;
}

export type CreateCryptoSessionRequestCurrency =
  (typeof CreateCryptoSessionRequestCurrency)[keyof typeof CreateCryptoSessionRequestCurrency];

export const CreateCryptoSessionRequestCurrency = {
  BTC: "BTC",
  ETH: "ETH",
  LTC: "LTC",
  USDT: "USDT",
  USDC: "USDC",
  SOL: "SOL",
} as const;

export interface CreateCryptoSessionRequest {
  /**
   * @minimum 1
   * @maximum 100
   */
  slotNumber: number;
  currency: CreateCryptoSessionRequestCurrency;
}

export interface CryptoSessionResponse {
  paymentId: string;
  address: string;
  amount: string;
  currency: string;
  expiresAt: string;
}

export interface VerifyCryptoRequest {
  paymentId: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export type DiscordCallbackParams = {
  code?: string;
  state?: string;
  error?: string;
};