# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Exe Joiner API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: slots
    description: Slot management
  - name: payments
    description: Payment operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/discord:
    get:
      operationId: discordLogin
      tags: [auth]
      summary: Initiate Discord OAuth2
      responses:
        "302":
          description: Redirect to Discord

  /auth/discord/callback:
    get:
      operationId: discordCallback
      tags: [auth]
      summary: Discord OAuth2 callback
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: error
          in: query
          required: false
          schema:
            type: string
      responses:
        "302":
          description: Redirect to dashboard or error page

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /slots:
    get:
      operationId: getSlots
      tags: [slots]
      summary: Get user slots
      responses:
        "200":
          description: User slots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SlotsResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/create-stripe-session:
    post:
      operationId: createStripeSession
      tags: [payments]
      summary: Create Stripe checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStripeSessionRequest"
      responses:
        "200":
          description: Stripe session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StripeSessionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/create-crypto-session:
    post:
      operationId: createCryptoSession
      tags: [payments]
      summary: Create crypto payment session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCryptoSessionRequest"
      responses:
        "200":
          description: Crypto session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CryptoSessionResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /payments/stripe-webhook:
    post:
      operationId: stripeWebhook
      tags: [payments]
      summary: Stripe webhook handler
      responses:
        "200":
          description: Webhook processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  /payments/verify-crypto:
    post:
      operationId: verifyCryptoPayment
      tags: [payments]
      summary: Verify crypto payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyCryptoRequest"
      responses:
        "200":
          description: Payment verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Payment not found or failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/settings:
    get:
      operationId: getAdminSettings
      tags: [admin]
      summary: Get admin settings
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: updateAdminSettings
      tags: [admin]
      summary: Update admin settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminSettingsUpdate"
      responses:
        "200":
          description: Settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSettings"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users with their slots
      responses:
        "200":
          description: All users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/users/{discordId}/slots:
    post:
      operationId: adminUpdateUserSlots
      tags: [admin]
      summary: Update a user's slot count or activate/deactivate slots
      parameters:
        - name: discordId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUpdateUserSlotsRequest"
      responses:
        "200":
          description: User slots updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    User:
      type: object
      properties:
        id:
          type: string
        discordId:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        isAdmin:
          type: boolean
      required:
        - id
        - discordId
        - username
        - isAdmin

    Slot:
      type: object
      properties:
        id:
          type: integer
        slotNumber:
          type: integer
        isActive:
          type: boolean
        purchasedAt:
          type: string
          nullable: true
        expiresAt:
          type: string
          nullable: true
        label:
          type: string
          nullable: true
      required:
        - id
        - slotNumber
        - isActive

    SlotsResponse:
      type: object
      properties:
        slots:
          type: array
          items:
            $ref: "#/components/schemas/Slot"
        totalSlots:
          type: integer
        pricePerDay:
          type: number
      required:
        - slots
        - totalSlots
        - pricePerDay

    AdminSettings:
      type: object
      properties:
        slotCount:
          type: integer
        pricePerDay:
          type: number
      required:
        - slotCount
        - pricePerDay

    AdminSettingsUpdate:
      type: object
      properties:
        slotCount:
          type: integer
          minimum: 1
          maximum: 100
        pricePerDay:
          type: number
          minimum: 0

    AdminUserEntry:
      type: object
      properties:
        discordId:
          type: string
        username:
          type: string
        avatar:
          type: string
          nullable: true
        activeSlots:
          type: integer
        totalSlots:
          type: integer
      required:
        - discordId
        - username
        - activeSlots
        - totalSlots

    AdminUsersResponse:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserEntry"
      required:
        - users

    AdminUpdateUserSlotsRequest:
      type: object
      properties:
        activeSlotCount:
          type: integer
          minimum: 0
          maximum: 100
      required:
        - activeSlotCount

    CreateStripeSessionRequest:
      type: object
      properties:
        slotNumber:
          type: integer
          minimum: 1
          maximum: 100
      required:
        - slotNumber

    StripeSessionResponse:
      type: object
      properties:
        sessionId:
          type: string
        url:
          type: string
      required:
        - sessionId
        - url

    CreateCryptoSessionRequest:
      type: object
      properties:
        slotNumber:
          type: integer
          minimum: 1
          maximum: 100
        currency:
          type: string
          enum: [BTC, ETH, LTC, USDT, USDC, SOL]
      required:
        - slotNumber
        - currency

    CryptoSessionResponse:
      type: object
      properties:
        paymentId:
          type: string
        address:
          type: string
        amount:
          type: string
        currency:
          type: string
        expiresAt:
          type: string
      required:
        - paymentId
        - address
        - amount
        - currency
        - expiresAt

    VerifyCryptoRequest:
      type: object
      properties:
        paymentId:
          type: string
      required:
        - paymentId

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error