# lib/api-zod/src/generated/types/adminUsersResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exe Joiner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminUserEntry } from "./adminUserEntry";

export interface AdminUsersResponse {
  users: AdminUserEntry[];
}