# lib/api-zod/src/generated/types/createCryptoSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Exe Joiner API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCryptoSessionRequestCurrency } from "./createCryptoSessionRequestCurrency";

export interface CreateCryptoSessionRequest {
  /**
   * @minimum 1
   * @maximum 100
   */
  slotNumber: number;
  currency: CreateCryptoSessionRequestCurrency;
}